requirements.txt
pyobjc-framework-Quartz==10.0
pyobjc-framework-ApplicationServices==10.0
